Fixes to heb-article, by Udi Fogiel (#12735)
For some reason theorems titles has \beginR without a \endR, and since
this is basically inside restricted horizontal mode, TeX complains about
a missing \endR.

I could not figure out why were they there in the first place so I
removed them and the output did not change, and the warnings are gone.

Beside that the numbering of the theorems is wrong (e.g. 1.0 instead of
0.1), the \@makelr thing should happen after babel is loaded.

The theorems layouts did not have a counter (in LyX) so I added one.